Active Ingredient

An Active Ingredient is any component that has medicinal properties, and supplies pharmacological activity or other direct effect in the diagnosis, cure, mitigation, treatment or prevention of disease, or to affect the structure or any function of the body of a human or an animal.

Usage

dpd_ai_id(ids)

dpd_ai_name(names)

dpd_ai_all()

Arguments

ids

Vector of drug product codes.

names

Vector of active ingredient names.

Value

A tibble with columns:

dosage_unit: Active ingredient dosage unit.

dosage_value: Quantitative value of the active ingredient dosage.

drug_code: Code assigned to each drug product.

ingredient_name: Ingredient name.

strength: Quantitative value of the active ingredient strength.

strength_unit: Active ingredient strength unit.

Examples

dpd_ai_id(48905)
#> # A tibble: 23 × 6
#>    dosage_unit dosage_value drug_code ingredient_name     strength strength_unit
#>    <chr>       <chr>            <int> <chr>               <chr>    <chr>        
#>  1 ""          ""               48905 VITAMIN A           1250     UNIT         
#>  2 ""          ""               48905 BETA-CAROTENE (PRO… 2916     UNIT         
#>  3 ""          ""               48905 VITAMIN B1          2.25     MG           
#>  4 ""          ""               48905 VITAMIN B2          3.75     MG           
#>  5 ""          ""               48905 NICOTINAMIDE        20       MG           
#>  6 ""          ""               48905 VITAMIN B6          5        MG           
#>  7 ""          ""               48905 D-PANTOTHENIC ACID  10       MG           
#>  8 ""          ""               48905 FOLIC ACID          0.2      MG           
#>  9 ""          ""               48905 BIOTIN              75       MCG          
#> 10 ""          ""               48905 VITAMIN B12         9        MCG          
#> # ℹ 13 more rows

dpd_ai_name('afatinib')
#> # A tibble: 12 × 6
#>    dosage_unit dosage_value drug_code ingredient_name     strength strength_unit
#>    <chr>       <chr>            <int> <chr>               <chr>    <chr>        
#>  1 ""          ""               90026 AFATINIB (AFATINIB… 20       MG           
#>  2 ""          ""               90027 AFATINIB (AFATINIB… 30       MG           
#>  3 ""          ""               90028 AFATINIB (AFATINIB… 40       MG           
#>  4 ""          ""              100203 AFATINIB (AFATINIB… 20       MG           
#>  5 ""          ""              100204 AFATINIB (AFATINIB… 30       MG           
#>  6 ""          ""              100205 AFATINIB (AFATINIB… 40       MG           
#>  7 ""          ""              100391 AFATINIB (AFATINIB… 20       MG           
#>  8 ""          ""              100392 AFATINIB (AFATINIB… 30       MG           
#>  9 ""          ""              100393 AFATINIB (AFATINIB… 40       MG           
#> 10 ""          ""              100635 AFATINIB (AFATINIB… 20       MG           
#> 11 ""          ""              100636 AFATINIB (AFATINIB… 30       MG           
#> 12 ""          ""              100637 AFATINIB (AFATINIB… 40       MG