Drug Product

Basic information about the product, such as brand name and Drug Identification Number.

Usage

dpd_drug_id(ids)

dpd_drug_din(dins)

dpd_drug_brand(brands)

dpd_drug_status(statuses)

dpd_drug_all()

dpd_drug_dins(names)

Arguments

ids

Drug product code.

dins

Drug Identification Number (DIN).

brands

Brand name.

statuses

Drug product status. See Details section.

names

Active ingredient name.

Value

A tibble with columns:

ai_group_no: Active Ingredient Group Number.

brand_name: Brand name for the drug product.

class_name: Therapeutic class of a drug product.

descriptor: Additional details or descriptions for a product, if applicable.

drug_code: Code assigned to each drug product.

drug_identification_number: Drug Identification Number.

number_of_ais: Number of Active Ingredients in a drug product.

company_name: Company name.

last_update_date: Date is updated any time certain, but not all, key fields are changed.

Details

status must be an integer, corresponding to the following statuses:

1: Approved

2: Marketed

3: Cancelled Pre Market

4: Cancelled Post Market

6: Dormant

9: Cancelled (Unreturned Annual)

10: Cancelled (Safety Issue)

Examples

dpd_drug_id(2049)
#> # A tibble: 1 × 9
#>   drug_code class_name drug_identification_number brand_name descriptor
#>       <int> <chr>      <chr>                      <chr>      <chr>     
#> 1      2049 Human      00326925                   SINEQUAN   ""        
#> # ℹ 4 more variables: number_of_ais <chr>, ai_group_no <chr>,
#> #   company_name <chr>, last_update_date <chr>

dpd_drug_din('00326925')
#> # A tibble: 1 × 9
#>   drug_code class_name drug_identification_number brand_name descriptor
#>       <int> <chr>      <chr>                      <chr>      <chr>     
#> 1      2049 Human      00326925                   SINEQUAN   ""        
#> # ℹ 4 more variables: number_of_ais <chr>, ai_group_no <chr>,
#> #   company_name <chr>, last_update_date <chr>

dpd_drug_status(10)
#> # A tibble: 1 × 9
#>   drug_code class_name drug_identification_number brand_name          descriptor
#>       <int> <chr>      <chr>                      <chr>               <chr>     
#> 1     15518 Human      02010593                   DRULA FADE CREAM S… ""        
#> # ℹ 4 more variables: number_of_ais <chr>, ai_group_no <chr>,
#> #   company_name <chr>, last_update_date <chr>