Route of Administration

Indicates the part of the body on which, through which or into which the product is to be introduced (e.g. oral, topical, intramuscular, rectal).

Usage

dpd_route(ids)

dpd_route_all()

Arguments

ids

Vector of drug product codes.

Value

A tibble with columns:

drug_code: Code assigned to each drug product.

route_of_administration_code: Code assigned to a route of administration.

route_of_administration_name: Route of administration.

Details

A product can have more than one route of administration (e.g. intravenous, intramuscular, intra-articular).

Examples

dpd_route(3)
#> # A tibble: 5 × 3
#>   drug_code route_of_administration_code route_of_administration_name
#>       <int>                        <int> <chr>                       
#> 1         3                           34 Intrapleural                
#> 2         3                           10 Intra-Articular             
#> 3         3                           33 Intraperitoneal             
#> 4         3                           49 Intravenous                 
#> 5         3                           31 Intrathecal