Product Status

The statuses listed below are a direct representation of the status available in DPD Online Query and the description of each status is summarized below.

Usage

dpd_status(ids)

Arguments

ids

Vector of drug product code.

Value

A tibble with columns:

drug_code: Code assigned to each drug product.

expiration_date: Latest expiry date of a product distributed in Canada, where the product has since been discontinued by the company.

external_status_code: Code assigned to an external drug's status.

history_date: Historical date associated with a drug's status.

lot_number: Latest lot number of a product distributed in Canada, where the product has since been discontinued by the company.

original_market_date: Original market date of a product.

Examples

dpd_status(10229)
#> # A tibble: 1 × 7
#>   drug_code status        history_date original_market_date external_status_code
#>       <int> <chr>         <chr>        <chr>                               <int>
#> 1     10229 Cancelled Po… 1997-10-15   1989-12-31                              4
#> # ℹ 2 more variables: expiration_date <chr>, lot_number <chr>